@@ -14,12 +14,12 @@
     "type":    "library",
     "license": "MIT",
     "require": {
-        "php": "^8.0",
+        "php": "^8.1",
         "illuminate/support": "^10.0"
     },
     "require-dev": {
         "ext-dom": "*",
-        "phpunit/phpunit": "^9.5.8"
+        "phpunit/phpunit": "^10.0"
     },
     "autoload": {
         "psr-4": {
