@@ -16,6 +16,5 @@ class Legend extends HtmlElement
      | -----------------------------------------------------------------
      */
 
-    /** @var  string */
-    protected $tag = 'legend';
+    protected string $tag = 'legend';
 }
