@@ -16,7 +16,7 @@ class Ul extends ListElement
      | -----------------------------------------------------------------
      */
 
-    protected $tag = 'ul';
+    protected string $tag = 'ul';
 
     /* -----------------------------------------------------------------
      |  Other Methods
@@ -25,16 +25,11 @@ class Ul extends ListElement
 
     /**
      * Make an item.
-     *
-     * @param  mixed $value
-     * @param  array $attributes
-     *
-     * @return \Arcanedev\Html\Elements\Element
      */
-    protected function makeItem($value, array $attributes)
+    protected function makeItem(mixed $value, array $attributes): HtmlElement
     {
-        return Element::withTag('li')
-                      ->attributes($attributes)
-                      ->html($value);
+        return static::withTag('li')
+            ->attributes($attributes)
+            ->html($value);
     }
 }
