@@ -4,12 +4,11 @@ declare(strict_types=1);
 
 namespace Arcanedev\Html\Exceptions;
 
+use Exception;
+
 /**
  * Class     InvalidChildException
  *
  * @author   ARCANEDEV <arcanedev.maroc@gmail.com>
  */
-class InvalidChildException extends \Exception
-{
-    //
-}
+class InvalidChildException extends Exception {}
