@@ -4,19 +4,21 @@ declare(strict_types=1);
 
 namespace Arcanedev\Html\Exceptions;
 
+use Exception;
+
 /**
  * Class     InvalidHtmlException
  *
  * @author   ARCANEDEV <arcanedev.maroc@gmail.com>
  */
-class InvalidHtmlException extends \Exception
+class InvalidHtmlException extends Exception
 {
     /* -----------------------------------------------------------------
      |  Main Methods
      | -----------------------------------------------------------------
      */
 
-    public static function onTag(string $tag)
+    public static function onTag(string $tag): static
     {
         return new static(
             "Can't set inner contents on `{$tag}` because it's a void element"
