@@ -4,12 +4,11 @@ declare(strict_types=1);
 
 namespace Arcanedev\Html\Exceptions;
 
+use Exception;
+
 /**
  * Class     VoidElementException
  *
  * @author   ARCANEDEV <arcanedev.maroc@gmail.com>
  */
-class VoidElementException extends \Exception
-{
-    //
-}
+class VoidElementException extends Exception {}
