@@ -21,12 +21,8 @@ trait AssertsHtmlStrings
 
     /**
      * Assert two Html strings are equals.
-     *
-     * @param  string                                         $expected
-     * @param  string|\Illuminate\Contracts\Support\Htmlable  $actual
-     * @param  string                                         $message
      */
-    public static function assertHtmlStringEqualsHtmlString(string $expected, $actual, string $message = ''): void
+    public static function assertHtmlStringEqualsHtmlString(string $expected, mixed $actual, string $message = ''): void
     {
         if ($actual instanceof Htmlable)
             $actual = $actual->toHtml();
@@ -45,12 +41,8 @@ trait AssertsHtmlStrings
 
     /**
      * Convert string to DOMDocument.
-     *
-     * @param  string  $html
-     *
-     * @return \DOMDocument
      */
-    protected static function convertToDomDocument($html): DOMDocument
+    protected static function convertToDomDocument(string $html): DOMDocument
     {
         return tap(new DOMDocument, function (DOMDocument $dom) use ($html) {
             $dom->loadHTML(preg_replace('/>\s+</', '><', $html));
